init obj

+++ bookmarks.php
@@ -1,5 +1,8 @@
 <?php
 
+        require('Item.php');
+
+
 define('DIR', '/bookmarks');
 /* define an array */
 function listExtensions(){
@@ -13,13 +16,32 @@ function listExtensions(){
  * @return string Html
  */
 function viewTree($dir=DIR) {
-    echo '<ul id="explorer">';
+
+    $citem = new Item(DIR, getcwd().$dir);
+    $citem->renderDir(true);
+
+    /*echo '<ul id="explorer">';
     foreach (getContentTree($dir) as $item) {
-        echo '<li>';
-        view($item);
+
+        $citem = new Item($item, getcwd().$dir);
+
+        
+        if ($citem->isDir()) {
+            $citem->renderDir();
+         } elseif ($citem->isFile()) {
+            echo '<li>';
+            $citem->renderFile();
+        //view($item);
         echo '</li>';
+            
+         };
+        
+
+        
     }
     echo '</ul>';
+
+    die;*/
 }
 
 /**
@@ -28,10 +50,12 @@ function viewTree($dir=DIR) {
  * @param string $dir name
  * @return array
  */
-function getContentTree($dir){
-    $path = getcwd().$dir;
+function getContentTree($dir=DIR){
+    if ($dir===DIR){
+        $dir = getcwd().$dir;
+    }
     $exclude_list = array(".", "..");
-    $items = array_diff(scandir($path), $exclude_list);
+    $items = array_diff(scandir($dir), $exclude_list);
     // exclude hidden files
     $items = array_filter($items, create_function('$a','return ($a[0]!=".");'));
     $items = sortItemsDir($items);
@@ -57,70 +81,6 @@ function sortItemsDir($items){
 }
 
 /**
-* View item
-*
-* @param string $item filename
-* @param string $dirPath relative path
-* @return string Html
-*/
-function view($item, $dirPath=null) {
-    // get path
-    if ($dirPath) {
-        $pathItem = $dirPath.'/'.$item;
-        $linkItem = str_replace(getcwd(), "", $pathItem);
-    } else {
-        $linkItem = DIR.'/'.$item;
-        $pathItem = getcwd().$linkItem;
-    }
-
-    // rendering
-    if (is_dir($pathItem)) {
-        renderDirName($item, $linkItem);
-        echo '<ul>';
-        foreach (getContentTree($linkItem) as $elm) {
-            echo '<li>';
-            view($elm, $pathItem);
-            echo '</li>';
-        }
-        echo '</ul>';
-    } else {        
-        if(strpos($item, '.mht') == false) {
-            $link = getUrl($item, $pathItem);
-        } else {
-            //$link = 'file:///'.$pathItem;
-            $link = $_SERVER["REQUEST_URI"].$linkItem;
-            /* var_dump($linkItem, '<br>');
-            var_dump($pathItem, '<br>');
-            var_dump($dirPath, '<br>');
-            var_dump($_SERVER["REQUEST_URI"], '<br>');*/
-        }
-        renderLink($item, $link);
-    }
-}
-
-/**
- * Render link
- *
- * @param string $item name
- * @param string $pathItem absolute
- * @return string Html
- */
-function renderLink($item, $pathItem){
-   $item =cleanFilename($item);
-    // rendering
-    echo '<a class="item" href="'.$pathItem.'" target="_blank" download="filename" /><span>'.$item.'</span></a>';
-}
-
-// extract url from file
-function getUrl($item, $pathItem){
-    foreach (file($pathItem) as $value) {
-        if(strpos($value, 'URL=') !== false){
-            return substr($value,4);
-        }
-    }
-}
-
-/**
  * Render Dir Name
  *
  * @param string $name
@@ -129,7 +89,6 @@ function getUrl($item, $pathItem){
  */
 function renderDirName($name, $linkItem){
     $arrayDepth = explode ('/', str_replace(DIR, "", $linkItem));
-
     $depth = sizeof($arrayDepth);
     echo '<h'.$depth.'>'.$name.'</h'.$depth.'>';
 }
@@ -147,7 +106,6 @@ function cleanFilename($name){
         }
         return $name;
     }
-    return;
 }
 
 ?>
